Inital Commit
These examples are a work in progress.  They are intended as a simple introduction on the basics of using the SE.R client.
The next things to do are finish the illustrations of using Streams and then extracting them into their own classes.
After that reader exercises will be added based on the simple interactions the code illustates.