Clay - Stream examples
Added examples of stream usage for SE.R.
This got more involved than planned due to the persistence of consumer groups and
the desire to have the code capable of executing consecutives runs without alteration.

Since I had to delete the consumer group at the end of execution,
it seemed like a bad idea not to explain what the code was and what it did.