Clay - Extracted pubsub and streams to separate classes
This is a cheap extraction of those examples into separate classes.  It
does a decent job of cleaning up the `Program` class but I'm wondering if
it might be better to separate out the consumer and producers into their
own classes as well.

Now would come more cleanup and the addition of exercises for the reader to finish.